SCRIPTS	new scripts and aliases
fuzzycd allows me to easily find any dir , either from a specificed
root, or / as default

mkcd makes a dir, then navigates into it. an alias dotsources the file
so that it happens within the same shell

tmux-session-template creates my typical workspace from a fuzzy search
for projects in ~/Projects or ~/templates. It can also be fed a dir